+++ src/react/views/LoginApp.js
@@ -7,12 +7,19 @@ import {
   View
 } from 'react-native'
 import Login from './Login'
-import { loginApp } from '../../style'
+import { loginApp, constants } from '../../style'
 import FacebookAuthWebView from './FacebookAuthWebView'
 import GoogleAuthWebView from './GoogleAuthWebView'
+import EmailPasswordAuthView from './EmailPasswordAuthView'
+import FancyTextButton from './input/FancyTextButton'
 import Backable from './Backable'
 
 export default class LoginApp extends Backable {
+
+  constructor (props) {
+    super(props)
+    this.state = {authFailed: false}
+  }
 
   get googleRoute () {
     return {
@@ -28,21 +35,38 @@ export default class LoginApp extends Backable {
     }
   }
 
+  get emailPasswordRoute () {
+    return {
+      index: 1,
+      Element: EmailPasswordAuthView
+    }
+  }
+
   renderScene ({index, Element}, navigator) {
     switch (index) {
       case 0:
         this.backAction = null
         return <Login
+          authFailed={this.state.authFailed}
           stateHandler={this.props.stateHandler}
+          onOpenEmailPasswordAuth={() => navigator.push(this.emailPasswordRoute)}
           onOpenGoogleAuth={() => navigator.push(this.googleRoute)}
           onOpenFacebookAuth={() => navigator.push(this.facebookRoute)}
         />
       case 1:
         this.backAction = () => navigator.pop()
-        return <Element
-          onSuccess={({secret, userId}) => this.props.stateHandler.updateAuth(userId, secret)}
-          onCancel={() => navigator.pop()}
-          onAuthFailed={() => navigator.pop()}/>
+        return <View style={{flex: 1}}>
+          <Element
+            stateHandler={this.props.stateHandler}
+            onSuccess={({secret, userId}) => this.props.stateHandler.updateAuth(userId, secret)}
+            onAuthFailed={() => {
+              navigator.pop()
+              this.setState({authFailed: true})
+            }}/>
+          <FancyTextButton
+            onPress={() => navigator.pop()} text='Cancel login'
+            style={{backgroundColor: constants.colorRed}}/>
+        </View>
       default:
         return null
     }
@@ -55,14 +79,6 @@ export default class LoginApp extends Backable {
       renderScene={(route, navigator) => <View style={loginApp.mainContainer}>
         {this.renderScene(route, navigator)}
       </View>}
-      navigationBar={<Navigator.NavigationBar
-        routeMapper={{
-          LeftButton: (route, navigator) => null,
-          Title: route => null,
-          RightButton: (route, navigator) => null
-        }}
-        style={loginApp.navigationBar}
-      />}
     />
   }
 }