+++ src/react/views/input/FancyFacebookButton.js
@@ -2,18 +2,20 @@
  * Created by budde on 25/02/2017.
  */
 import React from 'react'
-import FancyImageButton from './FancyImageButton'
+import FancyImageTextButton from './FancyImageTextButton'
 import { fancyFacebookButton } from '../../../style'
 
-const FancyFacebookButton = ({onPress, disabled}) => <FancyImageButton
+const FancyFacebookButton = ({text, onPress, disabled}) => <FancyImageTextButton
   style={fancyFacebookButton.button}
   onPress={onPress}
+  text={text}
   disabled={disabled}
   imageSource={require('../../../../img/facebook.png')}/>
 
 export default FancyFacebookButton
 
 FancyFacebookButton.propTypes = {
+  text: React.PropTypes.string.isRequired,
   onPress: React.PropTypes.func.isRequired,
   disabled: React.PropTypes.bool
 }